Refactor Conjur provider validations
