+++ pkg/provider/conjur/provider_test.go
@@ -68,77 +68,6 @@ func makeValidFindRef(search string, tags map[string]string) *esv1beta1.External
 	}
 }
 
-type ValidateStoreTestCase struct {
-	store *esv1beta1.SecretStore
-	err   error
-}
-
-func TestValidateStore(t *testing.T) {
-	testCases := []ValidateStoreTestCase{
-		{
-			store: makeAPIKeySecretStore(svcURL, svcUser, svcApikey, svcAccount),
-			err:   nil,
-		},
-		{
-			store: makeAPIKeySecretStore("", svcUser, svcApikey, svcAccount),
-			err:   fmt.Errorf("conjur URL cannot be empty"),
-		},
-		{
-			store: makeAPIKeySecretStore(svcURL, "", svcApikey, svcAccount),
-			err:   fmt.Errorf("missing Auth.Apikey.UserRef"),
-		},
-		{
-			store: makeAPIKeySecretStore(svcURL, svcUser, "", svcAccount),
-			err:   fmt.Errorf("missing Auth.Apikey.ApiKeyRef"),
-		},
-		{
-			store: makeAPIKeySecretStore(svcURL, svcUser, svcApikey, ""),
-			err:   fmt.Errorf("missing Auth.ApiKey.Account"),
-		},
-
-		{
-			store: makeJWTSecretStore(svcURL, "conjur", "", jwtAuthnService, "", "myconjuraccount"),
-			err:   nil,
-		},
-		{
-			store: makeJWTSecretStore(svcURL, "", jwtSecretName, jwtAuthnService, "", "myconjuraccount"),
-			err:   nil,
-		},
-		{
-			store: makeJWTSecretStore(svcURL, "conjur", "", jwtAuthnService, "", ""),
-			err:   fmt.Errorf("missing Auth.Jwt.Account"),
-		},
-		{
-			store: makeJWTSecretStore(svcURL, "conjur", "", "", "", "myconjuraccount"),
-			err:   fmt.Errorf("missing Auth.Jwt.ServiceID"),
-		},
-		{
-			store: makeJWTSecretStore("", "conjur", "", jwtAuthnService, "", "myconjuraccount"),
-			err:   fmt.Errorf("conjur URL cannot be empty"),
-		},
-		{
-			store: makeJWTSecretStore(svcURL, "", "", jwtAuthnService, "", "myconjuraccount"),
-			err:   fmt.Errorf("must specify Auth.Jwt.SecretRef or Auth.Jwt.ServiceAccountRef"),
-		},
-
-		{
-			store: makeNoAuthSecretStore(svcURL),
-			err:   fmt.Errorf("missing Auth.* configuration"),
-		},
-	}
-	p := Provider{}
-	for _, tc := range testCases {
-		_, err := p.ValidateStore(tc.store)
-		if tc.err != nil && err != nil && err.Error() != tc.err.Error() {
-			t.Errorf("test failed! want %v, got %v", tc.err, err)
-		} else if tc.err == nil && err != nil {
-			t.Errorf("want nil got err %v", err)
-		} else if tc.err != nil && err == nil {
-			t.Errorf("want err %v got nil", tc.err)
-		}
-	}
-}
-
 func TestGetSecret(t *testing.T) {
 	type args struct {
 		store      esv1beta1.GenericStore
